+++ hydromt_wflow/workflows/__init__.py
@@ -8,6 +8,7 @@ from hydromt_wflow.workflows.gauges import *
 from hydromt_wflow.workflows.glaciers import *
 from hydromt_wflow.workflows.grid import *
 from hydromt_wflow.workflows.landuse import *
+from hydromt_wflow.workflows.nbs import *
 from hydromt_wflow.workflows.reservoirs import *
 from hydromt_wflow.workflows.river import *
 from hydromt_wflow.workflows.rootzoneclim import *